+++ structurizr-dsl/src/main/java/com/structurizr/dsl/ElementsDslContext.java
@@ -27,7 +27,14 @@ class ElementsDslContext extends ModelItemsDslContext {
 
     @Override
     protected String[] getPermittedTokens() {
-        return new String[0];
+        return new String[] {
+                StructurizrDslTokens.RELATIONSHIP_TOKEN,
+                StructurizrDslTokens.TAG_TOKEN,
+                StructurizrDslTokens.TAGS_TOKEN,
+                StructurizrDslTokens.URL_TOKEN,
+                StructurizrDslTokens.PROPERTIES_TOKEN,
+                StructurizrDslTokens.PERSPECTIVES_TOKEN
+        };
     }
 
 }