+++ structurizr-dsl/src/main/java/com/structurizr/dsl/ModelItemParser.java
@@ -8,10 +8,6 @@ final class ModelItemParser extends AbstractParser {
 
     private final static int URL_INDEX = 1;
 
-    private final static int PERSPECTIVE_NAME_INDEX = 0;
-    private final static int PERSPECTIVE_DESCRIPTION_INDEX = 1;
-    private final static int PERSPECTIVE_VALUE_INDEX = 2;
-
     void parseTags(ModelItemDslContext context, Tokens tokens) {
         // tags <tags> [tags]
         if (!tokens.includes(TAGS_INDEX)) {
@@ -50,28 +46,6 @@ final class ModelItemParser extends AbstractParser {
 
         String url = tokens.get(URL_INDEX);
         context.getModelItem().setUrl(url);
-    }
-
-    void parsePerspective(ModelItemPerspectivesDslContext context, Tokens tokens) {
-        // <name> <description> [value]
-
-        if (tokens.hasMoreThan(PERSPECTIVE_VALUE_INDEX)) {
-            throw new RuntimeException("Too many tokens, expected: <name> <description> [value]");
-        }
-
-        if (!tokens.includes(PERSPECTIVE_DESCRIPTION_INDEX)) {
-            throw new RuntimeException("Expected: <name> <description> [value]");
-        }
-
-        String name = tokens.get(PERSPECTIVE_NAME_INDEX);
-        String description = tokens.get(PERSPECTIVE_DESCRIPTION_INDEX);
-        String value = "";
-
-        if (tokens.includes(PERSPECTIVE_VALUE_INDEX)) {
-            value = tokens.get(PERSPECTIVE_VALUE_INDEX);
-        }
-
-        context.getModelItem().addPerspective(name, description, value);
     }
 
 }