+++ structurizr-dsl/src/main/java/com/structurizr/dsl/ModelItemsParser.java
@@ -5,6 +5,7 @@ import com.structurizr.model.ModelItem;
 final class ModelItemsParser extends AbstractParser {
 
     private final static int TAGS_INDEX = 1;
+    private final static int URL_INDEX = 1;
 
     void parseTags(ModelItemsDslContext context, Tokens tokens) {
         // tags <tags> [tags]
@@ -18,6 +19,22 @@ final class ModelItemsParser extends AbstractParser {
             for (ModelItem modelItem : context.getModelItems()) {
                 modelItem.addTags(tags.split(","));
             }
+        }
+    }
+
+    void parseUrl(ModelItemsDslContext context, Tokens tokens) {
+        // url <url>
+        if (tokens.hasMoreThan(URL_INDEX)) {
+            throw new RuntimeException("Too many tokens, expected: url <url>");
+        }
+
+        if (!tokens.includes(URL_INDEX)) {
+            throw new RuntimeException("Expected: url <url>");
+        }
+
+        String url = tokens.get(URL_INDEX);
+        for (ModelItem modelItem : context.getModelItems()) {
+            modelItem.setUrl(url);
         }
     }